Tuesday 28/12 19:00
